% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhaustiveSearch.R
\name{breadthFirstSearch}
\alias{breadthFirstSearch}
\title{Exhaustive Search. Breadth First Search}
\usage{
breadthFirstSearch(data, class, featureSetEval)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{featureSetEval}{\itemize{
\item The measure for evaluate features
}}
}
\value{
A list is returned containing:
\describe{
\item{bestFeatures}{A vector with all features. Selected features are marked with 1, unselected features are marked with 0}
\item{bestFitness}{Evaluation measure obtained with the feature selection}
}
}
\description{
The method searches the whole features subset in breadth first order \insertCite{Kozen1992}{FSinR}
}
\examples{
## Breadth First Search for iris dataset (filter method)
breadthFirstSearch(iris, 'Species', binaryConsistency)
}
\references{
\insertAllCited{}
}
\author{
Adan M. Rodriguez

Francisco Aragón Royón
}
