% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2.R
\name{chiSquared}
\alias{chiSquared}
\title{Chi squared measure}
\usage{
chiSquared(data, class, features)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{features}{\itemize{
\item The feature or features to evalute individually
}}
}
\value{
\itemize{
\item The chi squared value for each selected feature
}
}
\description{
Calculates the Chi squared value \insertCite{Pearson1900}{FSinR}, evaluating the selected features individually
}
\examples{
chiSquared(iris,'Species','Sepal.Length')
}
\references{
\insertAllCited{}
}
