% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2.R
\name{cramer}
\alias{cramer}
\title{Cramer V measure}
\usage{
cramer(data, class, features)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{features}{\itemize{
\item The feature or features to evalute individually
}}
}
\value{
\itemize{
\item Cramer's V value for each selected feature
}
}
\description{
Calculates Cramer's V value \insertCite{Cramer1946}{FSinR}, evaluating features individually
}
\examples{
cramer(iris,'Species','Sepal.Length')
}
\references{
\insertAllCited{}
}
