% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhaustiveSearch.R
\name{deepFirstSearch}
\alias{deepFirstSearch}
\title{Exhaustive Search. Deep First Search}
\usage{
deepFirstSearch(data, class, featureSetEval)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{featureSetEval}{\itemize{
\item The measure for evaluate features
}}
}
\value{
A list is returned containing:
\describe{
\item{bestFeatures}{A vector with all features. Selected features are marked with 1, unselected features are marked with 0}
\item{bestFitness}{Evaluation measure obtained with the feature selection}
}
}
\description{
The method searches the whole features subset in deep first order \insertCite{Kozen1992}{FSinR}
}
\examples{
## Deep First Search for iris dataset (filter method)
deepFirstSearch(iris, 'Species', binaryConsistency)
}
\references{
\insertAllCited{}
}
\author{
Francisco Aragón Royón
}
