% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinationCoefficient.R
\name{determinationCoefficient}
\alias{determinationCoefficient}
\title{R Squared, to continous features}
\usage{
determinationCoefficient()
}
\value{
Returns a function that is used to generate an evaluation set measure using the R squared value for the selected features.
}
\description{
Generates an evaluation function that calculates the determinantion coefficient \insertCite{rsquared}{FSinR} of continuous features (set measure). This function is called internally within the \code{\link{filterEvaluator}} function.
}
\examples{
\dontrun{ 

## The direct application of this function is an advanced use that consists of using this 
# function directly to evaluate a set of features
## Classification problem

# Generate the evaluation function with Determination Coefficient
dc_evaluator <- determinationCoefficient()
# Evaluate the features (parameters: dataset, target variable and features)
dc_evaluator(longley, 'Employed', c('GNP', 'Population','Year'))
}
}
\references{
\insertAllCited{}
}
\author{
Adan M. Rodriguez
}
