% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Entropy}
\usage{
entropy(x)
}
\arguments{
\item{x}{\itemize{
\item Collection of values
}}
}
\value{
\itemize{
\item Entropy value
}
}
\description{
Calculates the entropy value, using the information theory.
}
\examples{
entropy(iris$Sepal.Length)
}
\author{
Adan M. Rodriguez

Alfonso Jiménez-Vílchez
}
