% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropyJ}
\alias{entropyJ}
\title{EntropyJ}
\usage{
entropyJ(x)
}
\arguments{
\item{x}{\itemize{
\item Collection of values
}}
}
\value{
\itemize{
\item EntropyJ value
}
}
\description{
Calculates the entropyJ value, using the information theory.
}
\examples{
entropyJ(iris$Sepal.Length)
}
\author{
Adan M. Rodriguez

Alfonso Jiménez-Vílchez
}
