% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fScore.R
\name{fscore}
\alias{fscore}
\title{F-score measure}
\usage{
fscore()
}
\value{
Returns a function that is used to generate an individual evaluation measure using the F-score.
}
\description{
Generates an evaluation function that calculates the F-score approach defined in \insertCite{Wang2018}{FSinR} (individual measure). This function is called internally within the \code{\link{filterEvaluator}} function.
}
\examples{
\dontrun{ 

## The direct application of this function is an advanced use that consists of using this 
# function directly to individually evaluate a set of features
## Classification problem

# Generate the evaluation function with F-Score
fscore_evaluator <- fscore()
# Evaluate the features (parameters: dataset, target variable and features)
fscore_evaluator(ToothGrowth, 'supp', c('len'))
}
}
\references{
\insertAllCited{}
}
