% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relief.R
\name{get.data.frame.from.formula}
\alias{get.data.frame.from.formula}
\title{get.data.frame.from.formula}
\usage{
get.data.frame.from.formula(formula, data)
}
\arguments{
\item{formula}{\itemize{
\item formula
}}

\item{data}{\itemize{
\item data
}}
}
\value{
\itemize{
\item data.frame
}
}
\description{
get.data.frame.from.formula
}
