% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{giniIndex}
\alias{giniIndex}
\title{Gini index measure}
\usage{
giniIndex(data, class, features)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{features}{\itemize{
\item The names of the selected feature
}}
}
\value{
\itemize{
\item The Gini index value for the selected features
}
}
\description{
This measure calculates the gini index \insertCite{Ceriani2012}{FSinR} of discrete features
}
\examples{
giniIndex(iris,'Species',c('Sepal.Width', 'Sepal.Length'))
}
\references{
\insertAllCited{}
}
\author{
Adan M. Rodriguez
}
