% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency.R
\name{IEConsistency}
\alias{IEConsistency}
\title{Inconsistent Examples consistency measure}
\usage{
IEConsistency(data, class, features)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples. Feature columns are expected to be factors, as all features should be discrete.
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{features}{\itemize{
\item The names of the selected features
}}
}
\value{
\itemize{
\item The consistency value for the selected features
}
}
\description{
Calculates the inconsistent examples consistency value \insertCite{DashLiu2003}{FSinR}, using hash tables
}
\examples{
IEConsistency(iris,'Species',c('Sepal.Width', 'Sepal.Length'))
}
\references{
\insertAllCited{}
}
\author{
Adan M. Rodriguez
}
