% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{is.discrete}
\alias{is.discrete}
\title{is.discrete(collection)}
\usage{
is.discrete(collection)
}
\arguments{
\item{collection}{\itemize{
\item A collection of values
}}
}
\value{
\itemize{
\item True if the collecion is discrete, False otherwise
}
}
\description{
Estimate if a collection contains discrete values
}
\examples{
is.discrete(mtcars$gear)
is.discrete(iris$Sepal.Length)
}
\author{
Alfonso Jiménez Vílchez
}
