% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{isDataframeDiscrete}
\alias{isDataframeDiscrete}
\title{isDataFrameDiscrete(dataframe)}
\usage{
isDataframeDiscrete(dataframe)
}
\arguments{
\item{dataframe}{\itemize{
\item A data frame
}}
}
\value{
\itemize{
\item True if all variables are discrete, False otherwise
}
}
\description{
Estimate if all variables in a data frame are discrete
}
\examples{
isDataframeDiscrete(mtcars)
isDataframeDiscrete(iris)
}
\author{
Alfonso Jiménez Vílchez
}
