% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jd.R
\name{Jd}
\alias{Jd}
\title{Jd evaluation measure}
\usage{
Jd(data, class, features)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{features}{\itemize{
\item The names of the selected features
}}
}
\value{
\itemize{
\item The value of the function for the selected features
}
}
\description{
Applies the discriminant function designed by Narendra and Fukunaga \insertCite{Narendra1977}{FSinR} to evaluate a set of features.
}
\examples{
Jd(ToothGrowth,'supp',c('len','dose'))
}
\references{
\insertAllCited{}
}
\author{
Alfonso Jiménez-Vílchez
}
