% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{normalization}
\alias{normalization}
\title{Normalize a data frame}
\usage{
normalization(data, class)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples
}}

\item{class}{\itemize{
\item The dependent variable
}}
}
\value{
\itemize{
\item The dataframe with the independent variables or features normalized
}
}
\description{
Takes in any data frame and normalize the data of their features
}
\details{
Normalize the data (without the class)
}
\examples{
normalization(iris,'Species')
}
\author{
Adan M. Rodriguez
}
