% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relief.R
\name{normalize.min.max}
\alias{normalize.min.max}
\title{normalize.min.max}
\usage{
normalize.min.max(data)
}
\arguments{
\item{data}{\itemize{
\item data
}}
}
\value{
\itemize{
\item normalized data
}
}
\description{
normalize.min.max
}
