% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relief.R
\name{relief}
\alias{relief}
\title{Relief}
\usage{
relief(data, class, features, neighbours.count = 5, sample.size = 10)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{features}{\itemize{
\item The feature or features to evalute individually
}}

\item{neighbours.count}{\itemize{
\item number of neighbours to find for every sampled instance
}}

\item{sample.size}{\itemize{
\item number of instances to sample
}}
}
\value{
\itemize{
\item a data.frame containing the worth of attributes in the first column and their names as row names
}
}
\description{
The relief algorithm \insertCite{Kira1992}{FSinR} finds weights of continous and discrete attributes basing on a distance between instances. Adapted from Piotr Romanski's Fselector package \insertCite{FSelectorPkg}{FSinR}.
}
\details{
relief
classification and regression
continous and discrete data
}
\examples{
relief(iris, 'Species', c('Sepal.Width', 'Sepal.Length'))
}
\references{
\insertAllCited{}
}
\author{
Alfonso Jiménez-Vílchez
}
