% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFSM.R
\name{RFSM}
\alias{RFSM}
\title{RFSM evaluation measure}
\usage{
RFSM(data, class, features, m = 5, k = 4)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{features}{\itemize{
\item The names of the selected features
}}

\item{m}{\itemize{
\item Number of iterations
}}

\item{k}{\itemize{
\item Number of neighbours
}}
}
\value{
\itemize{
\item The value of the function for the selected features
}
}
\description{
Feature set measure based on relief. Described in \insertCite{Arauzo2004}{FSinR}
}
\examples{
RFSM(iris, 'Species', c('Sepal.Width','Sepal.Length'))
}
\references{
\insertAllCited{}
}
\author{
Alfonso Jiménez-Vílchez
}
