% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequentialSelection.R
\name{sffs}
\alias{sffs}
\title{Sequential Floating Forward Selection}
\usage{
sffs(data, class, featureSetEval)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{featureSetEval}{\itemize{
\item The measure for evaluate features
}}
}
\value{
A list is returned containing:
\describe{
\item{bestFeatures}{A vector with all features. Selected features are marked with 1, unselected features are marked with 0}
\item{bestFitness}{Evaluation measure obtained with the feature selection}
}
}
\description{
The sffs method \insertCite{Pudil1994}{FSinR} starts with an empty set of features and add a single feature at each step with a view to improving the evaluation of the set. In addition, it checks whether removing any of the included features, improve the value of the set.
}
\examples{
## sffs method for mtcars dataset (filter method)
sffs(mtcars, 'mpg', mutualInformation)
}
\references{
\insertAllCited{}
}
\author{
Adan M. Rodriguez

Francisco Aragón Royón
}
