% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency.R
\name{binaryConsistency}
\alias{binaryConsistency}
\title{Binary consistency measure}
\usage{
binaryConsistency(data, class, features)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples. Feature columns are expected to be factors, as all features should be discrete.
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{features}{\itemize{
\item The names of the selected features
}}
}
\value{
\itemize{
\item The consistency value for the selected features
}
}
\description{
Calculates the binary consistency, also known as "Sufficiency test" from FOCUS \insertCite{AlmuallimDietterich1991}{FSinR}
}
\examples{
binaryConsistency(iris,'Species',c('Sepal.Width', 'Sepal.Length'))
}
\references{
\insertAllCited{}
}
\author{
Adan M. Rodriguez
}
