% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2.R
\name{cramer}
\alias{cramer}
\title{Cramer V measure}
\usage{
cramer()
}
\value{
Returns a function that is used to generate an individual evaluation measure using Cramer V.
}
\description{
Generates an evaluation function that calculates Cramer's V value \insertCite{Cramer1946}{FSinR}, evaluating features individually (individual measure). This function is called internally within the \code{\link{filterEvaluator}} function.
}
\examples{
\dontrun{ 

## The direct application of this function is an advanced use that consists of using this 
# function directly to individually evaluate a set of features
## Classification problem

# Generate the evaluation function with Cramer
cramer_evaluator <- cramer()
# Evaluate the features (parameters: dataset, target variable and features)
cramer_evaluator(iris,'Species',c('Sepal.Length'))
}
}
\references{
\insertAllCited{}
}
