% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinationCoefficient.R
\name{determinationCoefficient}
\alias{determinationCoefficient}
\title{R Squared, to continous features}
\usage{
determinationCoefficient(data, class, features)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{features}{\itemize{
\item The names of the selected features
}}
}
\value{
\itemize{
\item The R squared value for the selected features
}
}
\description{
This measure calculates the determinantion coefficient \insertCite{rsquared}{FSinR} of continuous features
}
\examples{
determinationCoefficient(iris,'Species',c('Sepal.Width', 'Sepal.Length'))
}
\references{
\insertAllCited{}
}
\author{
Adan M. Rodriguez
}
