% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fScore.R
\name{fscore}
\alias{fscore}
\title{F-score measure}
\usage{
fscore(data, class, features)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{features}{\itemize{
\item The name of the selected feature (only 1 feature)
}}
}
\value{
\itemize{
\item The value of the function for the selected feature
}
}
\description{
Evaluates a feature using the F-score approach defined in \insertCite{Wang2018}{FSinR}.
}
\examples{
fscore(ToothGrowth, 'supp', c('len'))
}
\references{
\insertAllCited{}
}
