% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{gainRatio}
\alias{gainRatio}
\title{The gain ratio measure}
\usage{
gainRatio(data, class, features)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples.  Feature columns are expected to be factors, as all features should be discrete.
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{features}{\itemize{
\item The names of the selected features
}}
}
\value{
\itemize{
\item The gain ratio value for the selected features.
}
}
\description{
This measure calculates the gain ratio value \insertCite{Quinlan1986}{FSinR}, using the information theory.
}
\examples{
gainRatio(iris,'Species',c('Sepal.Width', 'Sepal.Length'))
}
\references{
\insertAllCited{}
}
\author{
Adan M. Rodriguez
}
