% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlc.R
\name{MDLC}
\alias{MDLC}
\title{MDLC evaluation measure}
\usage{
MDLC(data, class, features)
}
\arguments{
\item{data}{\itemize{
\item A data frame with the features and the class of the examples
}}

\item{class}{\itemize{
\item The name of the dependent variable
}}

\item{features}{\itemize{
\item The names of the selected features
}}
}
\value{
\itemize{
\item MDLC value for the selected features
}
}
\description{
Applies the Minimum-Description_Length-Criterion (MDLC) \insertCite{Sheinvald1990}{FSinR} to evaluate a set of features.
}
\examples{
MDLC(iris,'Species',c('Sepal.Width', 'Sepal.Length'))
}
\references{
\insertAllCited{}
}
\author{
Alfonso Jiménez-Vílchez
}
